pyproject.toml
[project]
name = "user-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "mysql-connector-python>=9.5.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.45",
]
